requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argilla==1.25.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
effdet==0.4.1
emoji==2.10.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.2
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.31
layoutparser==0.3.4
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
opencv-python==4.6.0.66
openpyxl==3.1.2
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
protobuf==4.25.3
pyarrow==15.0.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
pypdfium2==4.28.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
streamlit==1.32.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.3.0
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.2.11
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.29.0
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.1
